Updated installer

+++ install.bash
@@ -1,9 +1,13 @@
 #!/bin/bash
-rm -rf LICENSE README.md
-pkg install python postgresql
-pip install -r requirements.txt
-rm -rf requirements.txt
-chmod +x main.py
-CAFETERIA="$(pwd)/scripts/main.py"
-echo "Anything is installed! GG."
-echo "Command to start the game - python $CAFETERIA"
+if [ "$(uname)" == "linux" ]; then
+  rm -rf LICENSE README.md
+  pkg install python postgresql
+  pip install -r requirements.txt
+  rm -rf requirements.txt
+  chmod +x main.py
+  CAFETERIA="$(pwd)/scripts/main.py"
+  echo "Anything is installed! GG."
+  echo "Command to start the game - python $CAFETERIA"
+else
+  echo "Installer can't starts on this system."
+fi
